+++ src/Components/AdSubmitForm/_adsubmitform.scss
@@ -2,7 +2,7 @@
 
 .type_focused{
   box-shadow: 0px 0px 15px -5px rgba(0, 0, 0, 0.68) !important;
-  border-top: 2px solid black !important;
+  // border-top: 2px solid black !important;
 }
 
 .ad-submit-main {
@@ -56,10 +56,13 @@
 
 //form
 .form-main {
-  .container {
-    border: 1px solid #c0c0c0;
+  // border: 1px solid #c0c0c0;
+  margin: 20px;
+  .container-fluid {
+    box-shadow: 0px 0px 15px -5px rgba(0, 0, 0, 0.68) !important;
     border-radius: 5px;
-    padding: 10px;
+    padding: 20px;
+    width: 78%;
   }
 
   .form-details {
@@ -69,13 +72,17 @@
       padding: 20px 0px;
       font-weight: 700;
       // text-decoration: underline;
-      text-align: left;
+      text-align: center;
       // letter-spacing: 0.15rem;
+    }
+
+    h3 {
+      padding-top: 20px;
     }
     label {
       display: block;
       font-size: $mid-font-size;
-      font-weight: 500;
+      font-weight: 600;
       color: #151515;
       text-align: left;
       padding: 10px 0px 0px 0px;
@@ -103,6 +110,26 @@
 
     input[type="checkbox"] {
       margin-right: 5px;
+    }
+
+    .room_button{
+      // width: 100%;
+      height: 40px;
+      border-radius: 5px;
+      font-size: $mid-font-size;
+      font-weight: 600;
+      transition: all 0.3s ease-in-out;
+      // padding: 15px;
+      // margin-top: 22px;
+      border: none;
+      color: white;
+      background-color: $primaryColor !important;
+    }
+
+    .room_button:hover{
+      background-color: #0e0e38 !important;
+      // color: black ;
+      // border: 1px solid black;
     }
 
     .ck-content {
